requirements.txt
Django==4.0.1
django-cleanup==5.2.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
Pillow==9.0.0
psycopg2==2.9.3
gunicorn